+++ doc/whats_new.rst
@@ -26,6 +26,13 @@ API
   with ``::`` instead of ``:``. This allow specifying URL channels.
   By `Thomas Moreau`_ (:gh:`758`)
 
+ENH
+---
+
+- Add a parallel backend system for ``benchopt run`` to setup distributed
+  run with ``dask`` and ``submitit``. See :ref:`parallel_run` for details.
+  By `Thomas Moreau`_ (:gh:`673`).
+
 FIX
 ---
 
@@ -258,7 +265,7 @@ CLI
   syntax. By `Tom Dupré la Tour`_ (:gh:`362`).
 
 - Add ``--slurm`` option in ``benchopt run`` to allow running the benchmark on
-  a SLURM cluster. See the :ref:`slurm_run` for more details on the config.
+  a SLURM cluster. See the :ref:`slurm_backend` for more details on the config.
   By `Thomas Moreau`_ (:gh:`407`)
 
 - Add ``benchopt archive`` to create a ``tar.gz`` archive with the benchmark's